refactor(#210): split game.cpp into per-concern files
Extract Game:: method implementations from the 1373-line game.cpp
into five focused sibling files following the existing tick/ pattern:

- game_init.cpp: constructor, destructor, Load(), SetupRenderDevice()
- game_render.cpp: Present(), surface resize, palette fade, TimerCallback()
- game_ui.cpp: Clay UI pipeline, screen stack push/pop/replace
- game_loop.cpp: Loop(), Tick(), GoToState(), GoBack()
- game_input.cpp: TickRumble(), TickGamepadMenuNav(), IsLiveMultiplayer()

game.cpp becomes a minimal stub with just the include list.
game.h unchanged. CMakeLists unchanged (uses GLOB_RECURSE).
No functionality changes.
