+++ clients/silencer/src/game/game_init.cpp
@@ -1,0 +1,323 @@
+#include "game.h"
+#include "config.h"
+#include "objecttypes.h"
+#include "state.h"
+#include "sdl3gpubackend.h"
+#include "tuibackend.h"
+#include <cstdlib>
+#include <cstring>
+#include <stdio.h>
+
+using namespace GameState;
+
+#ifndef SILENCER_LOBBY_HOST
+#define SILENCER_LOBBY_HOST "127.0.0.1"
+#endif
+#ifndef SILENCER_LOBBY_PORT
+#define SILENCER_LOBBY_PORT 517
+#endif
+#ifndef SILENCER_VERSION
+#define SILENCER_VERSION "00025"
+#endif
+#ifndef SILENCER_MAP_API_URL
+#define SILENCER_MAP_API_URL "http://127.0.0.1:8080"
+#endif
+
+Game::Game() : renderer(world), screenbuffer(640, 480),
+               uiClayService(uiClayBackend),
+               clientUi(uiClayService),
+               inGameUiController(world),
+               mapDownloader(world),
+               ambienceMixer(world, renderer, mapDownloader, fade_i),
+               screenContext(*this, world, renderer, world.lobby, keymap, updater, ambienceMixer, mapDownloader, window, renderdevice){
+	world.SetVersion(SILENCER_VERSION);
+	frames = 0;
+	fps = 0;
+	state = MAINMENU;
+	stateisnew = true;
+	fade_i = 0;
+	fadeStartMs = 0;
+	sharedstate = 0;
+	currentlobbygameid = 0;
+	lastannouncedgameid = 0;
+	lastannouncedstatus = 0;
+	joininggame = false;
+	memset(keystate, 0, sizeof(keystate));
+	gamepad = nullptr;
+	singleplayermessage = 0;
+	updatetitle = true;
+	minimized = false;
+	window = 0;
+	renderdevice = nullptr;
+	memset(palettecolors, 0, sizeof(palettecolors));
+	nextstateprocessed = false;
+#ifdef OUYA
+	quitscancode = SDL_SCANCODE_HOME;
+#else
+	quitscancode = SDL_SCANCODE_ESCAPE;
+#endif
+	chatEnterDebounce = false;
+	fullscreentoggled = false;
+	replayfile = 0;
+	controlPort = 0;
+	tuiInputPort = 0;
+	headless = false;
+	tui = false;
+	paused = false;
+	stepFramesRemaining = 0;
+	stepWallclockDeadlineMs = 0;
+}
+
+Game::~Game(){
+	// Join background download threads before tearing down SDL so they don't
+	// reference freed resources. The MapDownloader destructor would also do
+	// this, but it runs after SDL teardown — call it explicitly here.
+	mapDownloader.JoinAndShutdown();
+	// Bring the control server down first. Stop() runs the shutdown drain we
+	// registered at Load() time, fulfilling promises for both queued and
+	// pendingWaits commands so handler threads can unblock from fut.get() before
+	// we join them. Doing this before tearing down anything else keeps members
+	// pendingWaits/etc alive while the drain runs.
+	controlserver.Stop();
+	inputserver.Stop();
+	if(renderdevice){
+		renderdevice->Shutdown();
+		delete renderdevice;
+		renderdevice = nullptr;
+	}
+	if(window){
+		SDL_DestroyWindow(window);
+	}
+	world.resources.UnloadSounds();
+	Audio::GetInstance().Close();
+	MIX_Quit();
+	if(gamepad){ SDL_CloseGamepad(gamepad); gamepad = nullptr; }
+	SDL_Quit();
+}
+
+bool Game::Load(char * cmdline){
+	// CLI overrides for the lobby host / port — applied AFTER Config::Load
+	// below so the on-disk config doesn't clobber them. Empty strings / 0
+	// mean "no override; use the config value".
+	char lobbyHostOverride[256] = {0};
+	int  lobbyPortOverride = 0;
+	if((cmdline = strtok(cmdline, " "))){
+		do{
+			if(strncmp(cmdline, "-s", 2) == 0){ // dedicated server
+				setbuf(stdout, NULL);
+				setbuf(stderr, NULL);
+				char * lobbyaddress = strtok(NULL, " ");
+				char * lobbyport = strtok(NULL, " ");
+				char * gameid = strtok(NULL, " ");
+				char * accountid = strtok(NULL, " ");
+				char * gameport = strtok(NULL, " "); // optional: explicit UDP bind port (for Docker port mapping)
+				if(gameid && accountid && lobbyaddress && lobbyport){
+					unsigned short bindport = (gameport && atoi(gameport) > 0) ? (unsigned short)atoi(gameport) : 0;
+					world.Listen(bindport);
+					world.lobby.Connect(lobbyaddress, atoi(lobbyport));
+					do{
+						world.lobby.DoNetwork();
+						if(world.lobby.state == Lobby::RESOLVED){
+							world.lobby.Connect(lobbyaddress, atoi(lobbyport));
+						}
+						if(world.lobby.state == Lobby::DISCONNECTED || world.lobby.state == Lobby::CONNECTIONFAILED || world.lobby.state == Lobby::RESOLVEFAILED){
+							return false;
+						}
+						SDL_Delay(1);
+					}while(world.lobby.state != Lobby::CONNECTED);
+					world.gameplaystate = World::INLOBBY;
+					/*User * user;
+					do{
+						user = world.lobby.GetUserInfo(2);
+						world.lobby.DoNetwork();
+					}while(user->retrieving);
+					printf("name: %s, techslots: %d\n", user->name, user->agency[0].techslots);*/
+					world.dedicatedserver.Start(lobbyaddress, atoi(lobbyport), atoi(gameid), atoi(accountid));
+					char filename[256];
+					sprintf(filename, "replays/%d.zsr", atoi(gameid));
+					world.replay.BeginRecording(filename);
+					if(world.replay.IsRecording()){
+						world.replay.WriteHeader(world);
+					}
+					State * newstateobject = static_cast<State *>(world.CreateObject(ObjectTypes::STATE));
+					sharedstate = newstateobject->id;
+					newstateobject->state = 0;
+					state = NONE;
+				}
+			}else
+			if(strncmp(cmdline, "-r", 2) == 0){ // play replay
+				replayfile = strtok(NULL, " ");
+				GoToState(REPLAYGAME);
+			}
+			else if(strcmp(cmdline, "--control-port") == 0){
+				char * portstr = strtok(NULL, " ");
+				if(portstr){
+					controlPort = atoi(portstr);
+				}
+			}
+			else if(strcmp(cmdline, "--tui-input-port") == 0){
+				char * portstr = strtok(NULL, " ");
+				if(portstr){
+					tuiInputPort = atoi(portstr);
+				}
+			}
+			else if(strcmp(cmdline, "--headless") == 0){
+				headless = true;
+			}
+			else if(strcmp(cmdline, "--tui") == 0){
+				tui = true;
+			}
+			else if(strcmp(cmdline, "--lobby-host") == 0){
+				char * host = strtok(NULL, " ");
+				if(host){
+					strncpy(lobbyHostOverride, host, sizeof(lobbyHostOverride) - 1);
+					lobbyHostOverride[sizeof(lobbyHostOverride) - 1] = '\0';
+				}
+			}
+			else if(strcmp(cmdline, "--lobby-port") == 0){
+				char * portstr = strtok(NULL, " ");
+				if(portstr){
+					lobbyPortOverride = atoi(portstr);
+				}
+			}
+		}while((cmdline = strtok(0, " ")));
+	}
+	Config::GetInstance().Load();
+	if(lobbyHostOverride[0]){
+		strncpy(Config::GetInstance().lobbyhost, lobbyHostOverride, sizeof(Config::GetInstance().lobbyhost) - 1);
+		Config::GetInstance().lobbyhost[sizeof(Config::GetInstance().lobbyhost) - 1] = '\0';
+	}
+	if(lobbyPortOverride > 0){
+		Config::GetInstance().lobbyport = lobbyPortOverride;
+	}
+	LoadActiveKeymap(keymap);
+	if(world.dedicatedserver.active){
+		// Dedicated server: SDL3 always initialises the timer subsystem; no flags needed.
+		if(!SDL_Init(0)){
+			printf("Could not initialize SDL %s\n", SDL_GetError());
+			return false;
+		}
+	}
+	if(!world.dedicatedserver.active){
+		// SDL_INIT_GAMEPAD is opt-in; without it SDL_GetGamepads() returns
+		// nothing. Headless builds (CI / control-socket smoke tests) skip it
+		// since they don't need controller input. TUI mode keeps audio so the
+		// engine's normal mixer path works, but skips video and gamepad.
+		Uint32 sdlflags;
+		if(tui){
+			sdlflags = SDL_INIT_AUDIO;
+		}else if(headless){
+			sdlflags = 0;
+		}else{
+			sdlflags = SDL_INIT_VIDEO | SDL_INIT_AUDIO | SDL_INIT_GAMEPAD;
+		}
+		if(!SDL_Init(sdlflags)){
+			printf("Could not initialize SDL %s\n", SDL_GetError());
+			return false;
+		}
+		if(!headless && !tui) OpenFirstGamepad();
+		printf("Loading palette...\n");
+		if(!renderer.palette.SetPalette(0)){
+			return false;
+		}
+		if(tui){
+			if(!MIX_Init()){
+				printf("Could not initialize SDL_mixer: %s\n", SDL_GetError());
+			}
+			if(!Audio::GetInstance().Init(this)){
+				printf("Could not initialize audio\n");
+			}
+			Audio::GetInstance().SetMusicVolume(Config::GetInstance().musicvolume);
+			// No SDL_AddTimer (FPS counter title bar is window-only).
+			// No window — TUIBackend connects to the TS frontend over TCP.
+			if(!SetupRenderDevice()){
+				printf("Could not initialize TUI render device\n");
+				return false;
+			}
+			SetColors(renderer.palette.GetColors());
+		}else if(!headless){
+			if(!MIX_Init()){
+				printf("Could not initialize SDL_mixer: %s\n", SDL_GetError());
+			}
+			if(!Audio::GetInstance().Init(this)){
+				printf("Could not initialize audio\n");
+			}
+			Audio::GetInstance().SetMusicVolume(Config::GetInstance().musicvolume);
+			SDL_AddTimer(1000, TimerCallback, this);
+			//SDL_EnableUNICODE(true);
+			//SDL_EnableKeyRepeat(SDL_DEFAULT_REPEAT_DELAY, SDL_DEFAULT_REPEAT_INTERVAL);
+			//screen = SDL_SetVideoMode(640, 480, 8, SDL_DOUBLEBUF | SDL_SWSURFACE);
+			window = SDL_CreateWindow("Silencer", screenbuffer.w, screenbuffer.h,
+				SDL_WINDOW_RESIZABLE | SDL_WINDOW_HIGH_PIXEL_DENSITY |
+				(Config::GetInstance().fullscreen ? SDL_WINDOW_FULLSCREEN : 0));
+			SDL_StartTextInput(window);
+			SyncRenderSurfaceToWindowPixels();
+			if(!SetupRenderDevice()){
+				printf("Could not initialize GPU render device\n");
+				return false;
+			}
+			SetColors(renderer.palette.GetColors());
+			//SDL_Flip(screen);
+		}
+		// Headless mode skips SetColors() above, so palettecolors[] starts zeroed.
+		// It is first populated by SetColors() in the MAINMENU state handler in
+		// Loop(). Screenshot ops route to PostFrameReplies() AFTER Tick() runs,
+		// so the palette is always populated by the time a screenshot is captured.
+	}
+	printf("Loading resources...\n");
+	if(!world.resources.Load(this, world.dedicatedserver.active)){
+		printf("Could not load resources\n");
+		return false;
+	}
+	// GAS is loaded inside resources.Load() — rebuild buyable items now that
+	// GASLoader::Get().items is populated (World constructor ran too early).
+	world.LoadBuyableItems();
+	printf("Resources loaded\n");
+	lasttick = SDL_GetTicks();
+	RestartPaletteFade();
+	if(controlPort > 0){
+		auto drainPendingWaits = [this](){
+			for(auto& w : pendingWaits){
+				if(!w.cmd.reply) continue;
+				ControlReply rpl;
+				rpl.id = w.cmd.id;
+				rpl.ok = false;
+				rpl.code = "INTERNAL";
+				rpl.error = "server stopping";
+				w.cmd.reply->set_value(rpl);
+			}
+			pendingWaits.clear();
+		};
+		if(!controlserver.Start(controlPort, drainPendingWaits)){
+			fprintf(stderr, "[control] failed to start; continuing without\n");
+		}
+	}
+	if(tui && tuiInputPort > 0){
+		if(!inputserver.Start(tuiInputPort)){
+			fprintf(stderr, "[input] failed to start; continuing without\n");
+		}
+	}
+	return true;
+}
+
+bool Game::SetupRenderDevice(void){
+	if(tui){
+		TUIBackend *backend = new TUIBackend();
+		if(!backend->Init(nullptr)){
+			delete backend;
+			return false;
+		}
+		renderdevice = backend;
+		// TUIBackend ignores SetScaleFilter; safe to call.
+		renderdevice->SetScaleFilter(false);
+		return true;
+	}
+	SDL3GPUBackend *backend = new SDL3GPUBackend();
+	if(!backend->Init(window)){
+		delete backend;
+		return false;
+	}
+	renderdevice = backend;
+	renderdevice->SetScaleFilter(Config::GetInstance().scalefilter);
+	return true;
+}
